psxhw: Only call ResetIoCycle() if needed, mask HW IRQ bits
When HW IRQ status/mask ports are written to, it is only necessary to
force a sooner call to psxBranchTest() (by calling ResetIoCycle()) if
the end result of the write indicates a HW IRQ is pending.

Also, mask the values written to these ports so that games cannot
write outside the valid bitfield specified in PSX docs.

Removed unnecessary call to ResetIoCycle in psxHwReset().