mipsrec: Only call psxBranchTest() when necessary
It was being called after every single branch taken and at end of all blocks.
This amounted to tens of thousands of unnecessary calls per frame.

Now, psxRegs.cycle is compared to psxRegs.io_cycle_counter to determine when
psxBranchTest() needs to be called. This reduces the number of calls per frame
to often less than 30 or 40, a reduction of over 99%. Combined with the
simplications to psxBranchTest() itself introduced with new event queue,
CPU usage on GCW Zero 1GHz MIPS system is reduced by an average of 15%, and
an increase of 5-10FPS in many games that were not yet full speed.