mipsrec: Properly implement delayed read from COP2: fixes Tekken 2
This patch implements delayed read when MFC0, CFC0, MFC2, CFC2,
or LWC2 are placed in branch delay slot. If the branch is taken
the actual read to the general register happens after next opcode
is executed. If both opcodes (at branch target addr
and in delay slot) happen to use the same general register, the
result might be incorrect.

For now only BLTZ, BGTZ, BLTZAL, BGEZAL, BLEZ and BGEZ are
checked. This is enough to fix broken models in Tekken 2.

The trick: when branch is taken, the execution order of
problematic opcodes is reversed: first the one at destination
branch address, then the one in delay slot. After this, the
branch target address is altered to skip the first opcode of
the block.