gpu_unai: Sprite x,y sign-extend matches Mednafen, Rearmed gpu_neon
Fixes disappearing sprites in 1st level of SkullMonkeys as you walk right.

Nocash docs don't seem to mention a reason for this odd requirement.
I only found this fix from a very brief mention in the Mednafen
Changelog and then examining its source and gpu_neon's.

Applied this to both sprites (textured rects) and tiles (untextured rects).
Extending this behavior to polygons, as I would presume to be correct,
actually breaks poly rendering in many places. (So don't do that!)