Remove old SPU plugin 'spu_null'
It is not compatible with new spu_pcsxrearmed interface.

If need be, we can use Rearmed's 'spunull' adaptation, but better
to hold off on that until coming changes related to sound/frameskip
are completed. Even then, spunull of Rearmed isn't even compatible
with savestates that used a regular plugin, so is of little use.