mipsrec: Branch emitters load $ra in BD slot
Combined with the prior commit, this greatly reduces duplicate
loads of $ra and consequent load stalls in the recompiled code.

If block return address is not already loaded in $ra, load it in the
branch delay slot of the emitted branch instruction. This puts the
maximum distance between its load and usage. The branch-not-taken
codepath also often benefits by not having to load it separately.