mipsrec: Reduce load stalls in StoreToAddr()
Preload the first reg to be written in the store series while
doing initial address checks. This eliminates any stall before the
first actual store, and also reduces the psxRegs.writeok load stall.

Move the actual store instruction to a place below the code-invalidation
steps, to eliminate load stalls on the rest of most series.