mipsrec: Reduce load stalls in gen_SWL_SWR()
Similar to the prior commit:

Preload the first reg to be written in the store series while
doing initial address checks. This eliminates any stall before the
first actual store, and also reduces the psxRegs.writeok load stall.

Also eliminates duplicate load of first reg to be written in
indirect path (similar to prior commit but not mentioned there)