mipsrec: Blocks can return directly to dispatch loop
When HLE emulated BIOS is not in use, blocks can use faster direct return
jumps to the dispatch loop. Loads of $ra from stack are not needed.

Normally, blocks are executed by emu through two methods:
recExecute() and recExecuteBlock(). The latter is used to execute
until a certain PC is reached, and then returns. The former executes
forever, never returning. Both are used whether or not HLE emulated
BIOS is in use. HLE emulated BIOS can use both interchangeably during
execution, so the code generated cannot assume a fixed return address.

*However*, when using a standard BIOS, recExecuteBlock() is executed just
once, to maintain control over BIOS startup. Then, recExecute() is called
and remains in control of all execution, providing code with a fixed address
it can return to.

To take advantage of this fact, recExecute() is now a separate function that
is in charge of deciding which actual dispatch loop function to call. It will
detect when HLE BIOS is not in use, and clear the code cache before calling
the new direct-return dispatch loop. Code emitted from that point is faster.