mipsrec: Add new 'fastpath' block return method
When direct block return jumps are enabled (see prior commit), use
a faster version of dispatch loop when possible:

When a block branches backward to the top of itself, we know it's
been recompiled already. We now save its start code address before
entry, and allow it to return to a shorter, faster version of
dispatch loop that skips code table lookup and invalidation check.
This 'fastpath' takes only 10 cycles instead of 18. Analysis shows
that, on average, about 5% of in-game blocks executed take this
path. Overall, this means the emulator spends 2.5% less time inside
the dispatch loop. Furthermore, blocks no longer have to set $v0 to
a PC value before returning, as no code table lookup on it is needed.

Over 75 games have been tested using this method with no results showing
that the bypassing of code-invalidation-check causes issues.

This method could be extended to recExecuteIndirectReturn(), which would
allow it to be used when HLE BIOS is in use. However, it is best to
leave that alone for now as it already has existing compatibility issues.
It would also involve complicated tracking of which $ra value is
currently loaded inside a block. (Probably not worth it).