mipsrec: Optimize load of known-const spilled regs
In regcache.h, regMipsToHostHelper() blindly used LW to load a spilled
register from psxRegs.GPR[]. If the register value is known-const, this
caused unnecessary load stalls on subsequent reads of the reg.

Now, it checks if value is known-const and can be loaded with just one
ALU op. This eliminates any load stall while keeping code size the same.