mipsrec: Use new MIPS32r2 cache-flush instructions
Previously, we used Linux syscall cacheflush() to ensure recompiled
code becomes visible to the CPU. The syscall flushes the entire cache
that is specified. Instead, MIPS32r2 offers new unprivileged cache
instructions that allow much finer-grained control.

MIPS32R2_MakeCodeVisible authored and generously provided by
Nebuleon Fumika <nebuleon.fumika@gmail.com>, used with permission. (GPLv2)

Additionally, it was noticed that when we do need to call cacheflush()
on older non-MIPS32r2 platforms, it was only flushing the ICACHE. This is
fine, and should be faster, but it is now a define in the Makefile.gcw0.
If define 'DYNAREC_SKIP_DCACHE_FLUSH' is not present, cacheflush() will
flush *both* ICACHE & DCACHE. That is the most compatible method.