mipsrec: Eliminate load stalls in emitMFC2()
emitMFC2() matched the interpreter's C code, which worked fine, but
the C functions write back the register's returned value. Upon
examination of the GTE code, this seems to serve no purpose but to
create load stalls, and the recompiler no longer does this.

Also eliminated load stalls and branches when reg 28,29 are read.