Only allow reentry to frontend when cache status is normal.
If PS1 cache is isolated (psxRegs.writeok is false), it would likely cause
any savestates created to not work. This is simplest method of prevention.
Cache is only ever isolated for a very small amount of time, when a game
or the BIOS is doing an I-cache flush, so the risk is small but there.