mipsrec: Funcs w/ inline asm get 'noinline' attribute
Fixes bug in recExecute():

While testing unlikely cases in recExecute()'s decision of whether
direct return jumps are possible, a segfault crash occurred. When
indirect return jumps were to be used after forcing the test for
direct-jump-target range to fail, the subsequent calls to
recExecuteIndirectReturn_*() would cause an immediate crash.

It was discovered that GCC 4.9.1 was inlining the entire code for
recExecuteIndirectReturn_*() into recExecute(), their only callsites.

Adding 'noinline' function attribute to all functions containing inline
ASM fixed the problem. Ideally, these functions would be pure ASM ones
in a '.S' file. The inline-ASM-in-C-function method we've always used is
a hack, especially for what we need it to do. We'll keep things this way
for now, as it makes interfacing with C vars/defs easier and less brittle.

ALSO: When recExecute() is called, presumably after BIOS startup is done,
      it now flushes the code cache in all cases. Before, it only flushed
      when direct block return method was activated. Flushing in all cases
      instead should help reduce dead code leftover from the BIOS startup.