Makefile.gcw0: GCC skips div-by-zero checks for release builds
Normally, GCC emits checks for div-by-zero after every C/C++ division,
which would result in a runtime crash. This is only useful for development
or debugging, otherwise it just bloats code. Now that we have concept of
'dev' vs. 'release' builds in Makefile, it's OK to disable this.
This mainly would only help the 'gpu_unai' software rasterizer poly loops.