mipsrec: Smarter allocation of branch decision regs
MIPS branch decisions are made before any writeback from the BD slot.
Because we must emit the BD slot before emitting a branch, we
pessimistically asked for private allocations of the decision regs.
While this ensured the code emitted for BD slot did not affect the
branch, it increased reg pressure and forced unnecessary reg loads
and stalls.

Now, we only ask for private allocations when BD slot really writes
to a reg the branch reads.