mipsrec: Include loads in BD slot in load/store series
In a prior commit 47570460600872a94feb634f4ab36af95b930f0b
series of loads/stores could extend past a branch to include a store
in a BD slot. Now, we can also include loads in a BD slot. This allows
the BD slot load to share any overhead/checks with the series as a whole.