psxbios: Fix numerous issues with Digimon World on HLE Bios
On HLE BIOS, Digimon World would refuse to boot up.
The issue turned out to be an issue with strcat not being compliant with the official BIOS behaviour,
as well as an issue with EnterCritical not returning anything to v0.
This also properly addresses the crash issue in Medievil 2, as the experimental patch
was always returning 1, when it should only do so if all the bits were properly set.

Another issue it addresses is lack of input in Digimon World after booting up.

The issue turned out to be psxBios_PAD_init, which the game erroneously calls multiple times with invalid types.
This results in no input while in-game.
This was fixed by only allowing some types, as documented by nocash.